Move files to corresponding folder.
