+++ src/Conversion/IResultConverter.cs
@@ -1,4 +1,6 @@
-﻿namespace RaceResultConverter.Conversion;
+﻿using RaceResultConverter.DTO;
+
+namespace RaceResultConverter.Conversion;
 
 public interface IResultConverter<TFrom, TTo> where TFrom : IRaceResult where TTo : IRaceResult
 {
